dmm7510App/Db/dmm7510.proto: Changed stream device protocol reply timeout to 35 sec
Extended stream device protocol reply timeout to 35 sec in order to accommodate the worst case of communication delay, which is when the multimeter performs a measurement with nplc=15, filter count=100, and line synchronization=ON. After a software measure command, the reply takes roughly 26-27 seconds to come. The 8 seconds left are a large safety margin.