[iocBoot/iocdmm7510/, dmm7510App/Db/] Remove autosave for DMM option for now
The DMM7510 database does not support yet the use of the autosave tool. For this reason, the autosave PV set creation was removed from the stDMM7510.cmd file. However, the necessary files for using autosave with the DMM7510 database have been already included.

In order for the DMM database to support autosave, a feature is required that allows some of its PVs to be disabled depending on the configuration of other device parameters, e.g., the measurement function.