dcct.db: Modify download chain [In progress]
Remove settings unnecessary for new double-buffer acquisition
scheme from download chain and add links to new PVs (not created yet).

The main changes in this scheme, which reflect at the download
procedure, are that there is going to be a record to periodically
send software triggers, and that the 'Trigger Type' processing chain
will share part of its responsibilities with a new chain that Enables
acquisition. Therefore some links to records related to the trigger
type chain were removed from the download sequence, while the 'Enbl-Sel'
record was added.

The overall download sequence code arrangement also received an improvement,
for better readability.