dcct.db: Modify Trg Type proc chain [In Progress]
Simplify 'Trigger Type' selection processing chain
by removing unnecessary records to the new double-buffer
acquisition scheme.

The new 'Trigger Type' configuration sequence will now
only configure the triggers, but it will not start acquisition,
as it did previously.

The new record processing chain starting at the 'Enbl-Sel' record
(still to be created) will be responsible for assigning the correct
stimulus to the measurement. The triggers will be already working,
but not triggering any action until the acquisition is enabled.