dcct.db: Continue changes in Meas Trg chain
-Finish main changes in measurement trigger configuration
record processing chain.

-Move part of the configuration sequence to the processing
chain of a new record, 'Enbl-Sel'.

Most part of the acquisition logic is going to be in a
sequencer program, still to be added.
This is interesting because we need to keep track of which
buffer are we working with at each time, and to only send
messages related to that buffer in order to avoid 'drowning'
the device with SCPI commands. Problems resulting from sending
too many messages to the device have already been noticed,
sometimes making it unresponsive.