Make low limit detection only work during acquisition
Low Limit Detection is a feature of the DCCT IOC option
which configures the multimeter to send an electrical
pulse whenever the current measured by it drops bellow
a specified value. The pulse is sent approximately 1 us
the current drop is detected, when high frequency rejection
is not enabled.

This feature conflictis with the input level trigger
mode, in which the acquisition starts based on input level,
and there were inconsistensies in the way the PVs behaved.

Enabling Low Lim Detection would leave the 'MeasTrg-Sts' at
an 'Unknown' state, although everything would still work without
problems. Also, stopping acquisition (MeasTrg-Sel = None) would
set Low Lim Detection off.

Solution:

To solve the problems mentioned above, the Low Limit Detection
readback now only corresponds to part of the low limit configuration.
The limit detection is only truly enabled when acquisition is
enabled (this is probably the most intuitive), although its status
can indicate that it is configured if the other settings are correct.
The last setting, which is mapping the detection event to generate
an output trigger is left to be done by the sequencer program when
it starts acquisition.