Fix acquisition for less than 10 samples
Problem:

The current double buffer IOC version cannot read the DMM buffer correctly
when the number of samples per trigger is smaller than 10. This is due to
the smallest size configuration available for the DMM buffers. Since currently
each trigger must completely fill the buffer in order for the sequencer to detect
new readings, sample counts smaller than 10 are not possible.

Solution:

A simple way to solve the problem is to monitor the position of the buffer last reading.
The IOC starts reading the buffer from the new readings initial position until its last.

The buffer size is set to a multiple of the sample count to simplify the logic.