dcct.db: Add monitor for time between measurements
Add the PV 'MeasUpdatePeriod-Mon', which monitors the difference
between timestamps of two consecutive current measurements.

This PV aims to provide an estimate of the true interval between
measurements, as opposed to the set point 'measurement period' set
by the operator, which is the device parameter controlling the acquisition
duration (for the given number of samples). The readback value of the
'measurement period' only says when the parameter was successfully set
in the device, but this setting cannot guarantee the exact period
considering IOC and multimeter processing time and the communication
between both.

This PV ignores the first measurement after an acquisition start,
since there is no previous valid measurement for it to compare
timestamps.