dcct.db: Add PV to refresh history of current
Add a PV that allows clearing the current history buffer.

This is useful when the current meaurement range is changed or
when the measurement is restarted after a relatively long period.
Clearing the buffer allows a graph displaying the current history
to show only the new data (i.e., the data after the refresh).