.../sncDCCT.stt: Add failure recovery procedure
Add a procedure in the 'sncDCCT.stt' program state machine to
allow the IOC to automatically recover from a power fault or
network problem during an acquisition.

When a problem happens during acquisition, the IOC stops any
action and waits until network connection is reestablished.
At this point the IOC downloads all parameters to the instrument
and restarts acquisition with previous configurations.

When a problem happens during "idle" time, the IOC detects a
network problem after the connection timeout and wait until
reconnection to download parameters. If an acquisition command
is received during this period when the connection is down,
the IOC ignores the command.