dmm7510App/Db/dcct.db: Fix MeasTrg-Sel config
Change 'MeasTrg-Sel' record OOPT (calcout output condition) field from "On Change" to always.

Before, this PV only triggered the update of low level parameters when its value changed.
Now it causes updates whenever it processes. This is important because otherwise a
download operation is not able to resend trigger parameters to the instrument if the
trigger option has not changed, and this might be necessary after a power failure or
some other problem like this happens.