*App/Db/dcct.db: Improve download reliability
Improve reliability of download finished status.

Before there was no guarantee that the commands sent tooo the instrument
were accepted before considering the download finished.

Now, the 'MeasTrg-Sts' PV (measurement trigger type) must have been
configured for the download status ('DownloadDone-Mon' PV) to be 'Finished'.

It is specially important to monitor the triggr type PV because it configures different low
level parameters and its readback is updated by reading the instrument configured
values. So, this action should take some time to finish. Since it is the last
parameter to be configured during download (with exception of the enable PV), after
it is finished, other parameters should also be configured since the instrument receives
the configuration commands for them before.