[DCCT]: Fix MeasTrg-Sts state calc
The calculation of the value of the 'MeasTrg-Sts' record had a mistake.
For the fast acquisition mode, instead of taking into consideration the value of the
multimeter 'DATrEdgeSlp-Sts' PV (Digitizer Analog Trigger Edge Slope status),
it was monitoring the 'MATrEdgeSlp-Sts' PV (Measure Analog Trigger Edge Slope status),
which is related to the normal acquisition mode.

The monitored PV has been fixed.