[DCCT] Add PV to abort acquisition
From the perspective of the Sirius machine high level application,
it is interesting to have a separate Abort command, in addition of
being able to set the PV Enbl-Sel to 0.

The advantage is that with the new PV, the current high level
machine configuration service can stop DCCT acquisition, apply
configuration changes and then proceed with acquisition, while
using only one setup file.