[DCCT] Add new bit to realiable meas status
Add a 4th bit (in the most significative bit position
compared to the others, i.e., to their left), which
provides information when the measured current exceeds
the configured range.

When the last measurement is greater than the range configured
value, this bit goes to 1. If the measurement returns within
accepted values, the bit goes to 0.

It is important to keep measurement values inside the range
because, although the measured values could keep accurate
for some margin outside the range, the toroid sensor could
saturate at some point.

If the sensor saturates, measurements are not realiable anymore.
In order to desaturate it, it is necessary to power cycle the
electronics.