[DCCT] Update OutOfRange PV when range change
Update OutOfRange-Mon PV also when range changes.

This is useful to avoid the status bit being in warning
state right after the IOC is started.

Since the IOC may start in an invalid range configuration,
the first processing of the OutOfRange-Mon PV could leave it
in warning until a measurement was started, even if the range
was automatically corrected. This behavior is fixed by this commit.